# lib/api-client-react/src/dashboard/models/createProject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * P1Dashboard
 * Shared field, scheduling and commercial inbox contract. Other office routes remain documented in docs/dashboard/API.md.
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProjectPhasesItem } from './createProjectPhasesItem';

export interface CreateProject {
  /**
     * @minItems 1
     * @maxItems 100
     */
  propertyIds: string[];
  /**
     * @minLength 1
     * @maxLength 10000
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 10000
     */
  scope: string;
  /** @maxItems 50 */
  phases?: CreateProjectPhasesItem[];
}