# lib/api-client-react/src/dashboard/models/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * P1Dashboard
 * Shared field, scheduling and commercial inbox contract. Other office routes remain documented in docs/dashboard/API.md.
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectPhasesItem } from './projectPhasesItem';

/**
 * Staff-only multi-client delivery container. Jobs, estimates, agreements and billing stay property-specific.
 */
export interface Project {
  id: string;
  name: string;
  scope: string;
  phases?: ProjectPhasesItem[];
  client_names?: string;
  property_names?: string;
  created_at?: string;
}