# lib/api-spec/dashboard.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "P1Dashboard",
    "version": "0.1.0",
    "description": "Shared field, scheduling and commercial inbox contract. Other office routes remain documented in docs/dashboard/API.md."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/projects": {
      "get": {
        "operationId": "listProjects",
        "description": "Staff-only Project containers. The response identifies participating clients and properties but is never a client portal projection.",
        "responses": { "200": { "description": "Shared Projects", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } } } } }, "401": { "description": "Verified session required" }, "403": { "description": "Dispatch role required" } }
      },
      "post": {
        "operationId": "createProject",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProject" } } } },
        "responses": { "201": { "description": "Project created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedResource" } } } }, "400": { "description": "Participating properties invalid" }, "401": { "description": "Verified session required" }, "403": { "description": "Manager role required" } }
      }
    },
...
    "/inspections": {
      "get": {
        "operationId": "listInspectionReports",
        "description": "Owner, manager, and dispatch receive operational inspection reports. Clients receive only explicitly published reports for accessible properties and never the submitting staff user ID.",
        "responses": { "200": { "description": "Authorized reports", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/InspectionReport" } } } } }, "401": { "description": "Verified session required" }, "403": { "description": "Role or required-account MFA denied" } }
      },
      "post": {
        "operationId": "createInspectionReport",
        "description": "Records an operational inspection report. It remains private until an owner or manager explicitly publishes it.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateInspectionReport" } } } },
        "responses": {
          "201": { "description": "Inspection created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedResource" } } } },
          "400": { "description": "Invalid report" },
          "401": { "description": "Verified session required" },
          "403": { "description": "Role or required-account MFA denied" },
          "404": { "description": "Operational property not accessible" }
        }
      }
    },
    "/inspections/{id}/publish": {
      "post": {
        "operationId": "publishInspectionReport",
        "description": "Owner or manager only. Makes a report client-visible; it does not alter work, files, billing, payment, or underlying property condition data. Repeating publication is safe.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Published or already published", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PublicationReceipt" } } } }, "401": { "description": "Verified session required" }, "403": { "description": "Owner or manager role required" }, "404": { "description": "Operational inspection not found" } }
      }
    },
...
    "/recurring-services": {
      "get": {
        "operationId": "listAgreementRecurrences",
        "responses": {
          "200": {
            "description": "Authorized response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgreementRecurrenceOption"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Verified session required"
          },
          "403": {
            "description": "Role or required-account MFA denied"
          },
          "404": {
            "description": "Record inaccessible"
          },
          "409": {
            "description": "Stale version, eligibility or approved-cap conflict; retain edits for review"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProject": {
        "type": "object",
        "additionalProperties": false,
        "required": ["propertyIds", "name", "scope"],
        "properties": {
          "propertyIds": { "type": "array", "minItems": 1, "maxItems": 100, "items": { "type": "string", "format": "uuid" } },
          "name": { "type": "string", "minLength": 1, "maxLength": 10000 },
          "scope": { "type": "string", "minLength": 1, "maxLength": 10000 },
          "phases": { "type": "array", "maxItems": 50, "items": { "type": "object", "required": ["name", "complete"], "properties": { "name": { "type": "string" }, "complete": { "type": "boolean" } } } }
        }
      },
      "Project": {
        "type": "object",
        "description": "Staff-only multi-client delivery container. Jobs, estimates, agreements and billing stay property-specific.",
        "required": ["id", "name", "scope"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "scope": { "type": "string" },
          "phases": { "type": "array", "items": { "type": "object" } },
          "client_names": { "type": "string" },
          "property_names": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreatedResource": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateInspectionReport": {
        "type": "object",
        "additionalProperties": false,
        "required": ["propertyId", "title", "findings"],
        "properties": {
          "propertyId": { "type": "string", "format": "uuid" },
          "title": { "type": "string", "minLength": 1, "maxLength": 10000 },
          "findings": { "type": "array", "maxItems": 100, "items": { "$ref": "#/components/schemas/InspectionFinding" } }
        }
      },
...
      "InspectionFinding": {
        "type": "object",
        "additionalProperties": false,
        "required": ["label", "condition", "note"],
        "properties": {
          "label": { "type": "string", "minLength": 1, "maxLength": 200 },
          "condition": { "type": "string", "enum": ["good", "monitor", "action_required", "not_assessed"] },
          "note": { "type": "string", "maxLength": 5000 }
        }
      },
      "InspectionReport": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "property_id", "title", "findings", "published", "created_at", "property_name"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "property_id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "description": "Submitting staff identity; omitted from client responses." },
          "title": { "type": "string" },
          "findings": { "type": "array", "items": { "$ref": "#/components/schemas/InspectionFinding" } },
          "published": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "property_name": { "type": "string" }
        }
      },
      "PublicationReceipt": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ok"],
        "properties": { "ok": { "type": "boolean" } }
      },
...
      "AgreementRecurrenceOption": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "id",
          "property_id",
          "title",
          "billing_mode",
          "property_name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "property_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "billing_mode": {
            "type": "string",
            "enum": ["fixed_monthly", "per_visit"]
          },
          "property_name": {
            "type": "string"
          }
        },
        "description": "Selected fields from the existing operational snapshot; other persisted fields may be present."
      },
...
    }
  }
}